Adding the posibility of resetting som variables
As I needed some variables with same name and purpose in several sections and had also only some of these variables set and others only used in a default mode (e.g. "lineEnding" has normally to be an empty string, but in some cases, I needed a ";" (semicolon), and to avoid to repeat "thousand times" the line --lineEndig=""--, I came up with the reset idea.
In the ini-file das to be a "reset" section, in which all values are declared, which are often used and mostly with a default value. So only in some section where this varys, one must the define a specific value - in all other cases, the value defined in the "reset" exction will fit. The "reset" command will be called as first command in every section function, if no "reset" section exists, the comman quits itself quietly.